render descriptions/comments with a custom filter
If the `RICHTEXT_FILTER` setting is set to a path to a function,
it is used to render descriptions and comments.

TextFields that must be rendered must have the `richtext`
attribute set to True.

By default, if the settings is not set, it just breaks paragraph (linebreaks).

A markdown filter (`'openPLM.plmapp.filters.markdown_filter'`) is available
(requires python-markdown). Specific extensions (links, ...) will come later.

If the `RICHTEXT_PLAIN_FILTER` setting is set, it is used to render indexed
field (to remove garbage characters). By default, it returns the same string.
If only `RICHTEXT_FILTER` is set, it returns a clean version (stripped tags)
of the html rendered content.

Custom editors (to replace a simple textarea) are not yet supported.